+++ test/dom_tools_test.dart
@@ -101,5 +101,161 @@ void main() {
       expect(await state1.getAsync("b"), equals({'x': 10, 'y': 'Y'}));
       expect(state1.get("b"), equals({'x': 10, 'y': 'Y'}));
     });
+
+    test('extensions', () {
+      var div = DivElement()..setInnerHtml('''
+      
+        <a id="lnk1" href="link1.html">link 1</a>
+        <a id="lnk2" href="link2.html#h2" class="hash hash2">link 2</a>
+        <a id="lnk3" href="link3.html#h3" class="hash">link 3</a>
+        
+        <input id="chk1" type="checkbox" value="a" checked>
+        <input id="chk2" type="checkbox" value="b" class="chkb">
+        
+        <input id="rad1" type="radio" value="c" checked>
+        <input id="rad2" type="radio" value="d">
+        
+        <img id="img1" src='foo.png'>
+        
+      ''');
+
+      var innerElements = _elementsToTag(div.children);
+
+      expect(
+          innerElements,
+          equals([
+            '<a id=lnk1 href=link1.html>',
+            '<a id=lnk2 href=link2.html#h2 class=hash+hash2>',
+            '<a id=lnk3 href=link3.html#h3 class=hash>',
+            '<input id=chk1 type=checkbox value=a checked>',
+            '<input id=chk2 type=checkbox value=b class=chkb>',
+            '<input id=rad1 type=radio value=c checked>',
+            '<input id=rad2 type=radio value=d>',
+            '<img id=img1 src=foo.png>',
+          ]));
+
+      {
+        expect(
+            _elementsToTag(div.children.whithClass('hash')),
+            equals([
+              '<a id=lnk2 href=link2.html#h2 class=hash+hash2>',
+              '<a id=lnk3 href=link3.html#h3 class=hash>',
+            ]));
+
+        expect(_elementsToTag(div.children.whithClass('hash2')),
+            equals(['<a id=lnk2 href=link2.html#h2 class=hash+hash2>']));
+
+        expect(_elementsToTag(div.children.whithClass('chkb')),
+            equals(['<input id=chk2 type=checkbox value=b class=chkb>']));
+
+        expect(
+            _elementsToTag(div.children.whithClasses(['hash', 'hash2'])),
+            equals([
+              '<a id=lnk2 href=link2.html#h2 class=hash+hash2>',
+            ]));
+      }
+
+      {
+        expect(_elementsToTag(div.children.withID('lnk2')),
+            equals(['<a id=lnk2 href=link2.html#h2 class=hash+hash2>']));
+
+        expect(_elementsToTag(div.children.withID('img1')),
+            equals(['<img id=img1 src=foo.png>']));
+      }
+
+      {
+        var sel = div.selectAnchorElements();
+
+        print(_elementsToTag(sel));
+
+        expect(sel.length, equals(3));
+
+        expect(
+            _elementsToTag(sel),
+            equals([
+              '<a id=lnk1 href=link1.html>',
+              '<a id=lnk2 href=link2.html#h2 class=hash+hash2>',
+              '<a id=lnk3 href=link3.html#h3 class=hash>'
+            ]));
+
+        var selLinks = div.selectAnchorLinks();
+
+        expect(selLinks.map((e) => e.split('/').last),
+            equals(['link1.html', 'link2.html#h2', 'link3.html#h3']));
+
+        var selLinksTargets = div.selectAnchorLinksTargets();
+
+        expect(selLinksTargets, equals(['h2', 'h3']));
+      }
+
+      {
+        var sel = div.selectCheckboxInputElement();
+
+        print(_elementsToTag(sel));
+
+        expect(sel.length, equals(2));
+
+        expect(
+            _elementsToTag(sel),
+            equals([
+              '<input id=chk1 type=checkbox value=a checked>',
+              '<input id=chk2 type=checkbox value=b class=chkb>'
+            ]));
+
+        expect(sel[0].id, equals('chk1'));
+        expect(sel[0].value, equals('a'));
+        expect(sel[0].checked, isTrue);
+
+        expect(sel[1].id, equals('chk2'));
+        expect(sel[1].value, equals('b'));
+        expect(sel[1].checked ?? false, isFalse);
+      }
+
+      {
+        var sel = div.selectRadioButtonInputElement();
+
+        print(_elementsToTag(sel));
+
+        expect(sel.length, equals(2));
+
+        expect(
+            _elementsToTag(sel),
+            equals([
+              '<input id=rad1 type=radio value=c checked>',
+              '<input id=rad2 type=radio value=d>'
+            ]));
+
+        expect(sel[0].id, equals('rad1'));
+        expect(sel[0].value, equals('c'));
+        expect(sel[0].checked, isTrue);
+
+        expect(sel[1].id, equals('rad2'));
+        expect(sel[1].value, equals('d'));
+        expect(sel[1].checked ?? false, isFalse);
+      }
+
+      {
+        var sel = div.selectImageElement();
+
+        print(_elementsToTag(sel));
+
+        expect(sel.length, equals(1));
+
+        expect(_elementsToTag(sel), equals(['<img id=img1 src=foo.png>']));
+
+        expect(sel[0].id, equals('img1'));
+        expect(sel[0].src, endsWith('foo.png'));
+      }
+    });
   });
+}
+
+List<String> _elementsToTag(Iterable<Element> children) {
+  return children
+      .map((e) => '<${e.tagName.toLowerCase()} ${e.attributes.entries.map((e) {
+            var key = e.key;
+            var value = e.value.replaceAll(' ', '+');
+            return value.isNotEmpty ? '$key=$value' : key;
+          }).join(' ')}>')
+      .toList();
 }